Changed the way hours for resources are calculated.
Instead of only counting the hours written on tickets that have been created today, I now calculate all time entries of today, independent of the tickets.

This should solve the issues a lot of users are having with the numbers appearing to be off, since they compare them with the timesheets in Autotask (which is a logical thing to do).