These changes are all made to speed up the queries. They are made on the idea that all queries should only fetch what they need, no attached (unnecessary) models.
The error handling has also changed. Where I would set a boolean to see
if any errors occured, now the error handling gets done through Exceptions.

A lot of the logic in the ImportFromAutotaskShell has been moved to seperate
function (kept in the same file) to improve readability and debugging.